bugfix: select commands keep anchor position
Both `jump-and-select.jump{Forward,Backward}Select` and `jump-and-select.jump{Forward,Backward}SelectMultiMode` commands set selection anchors as the current cursor position (selection.active position). This results in a bug where a current selection is overridden by these commands, which was particularly obvious with multimode creating a new selection for every jump made. This bugfix changes the commands to keep the original anchor position so that jump-and-select can be used to extend selections.